requirements.txt
blacksheep >= 2.1.0
uvicorn
uvloop; sys_platform != 'win32'
pyjwt
piccolo[postgres]
argon2-cffi
uuid_utils
python-dotenv
msgspec[toml]
cacheout
pyjwt[crypto]
colorlog